修改手续费为小数

+++ app/src/main/java/com/jnhyxx/html5/domain/order/FuturesFinancing.java
@@ -141,7 +141,7 @@ public class FuturesFinancing {
     public static class AssetsBean {
         private double marginBeat;
         private Map<String, Double> stopWinBeats;
-        private int fees;
+        private double fees;
         private Map<String, Double> marginBeatHands;
         private int isDefault;
         private Map<String, Double> feesHands;
@@ -169,11 +169,11 @@ public class FuturesFinancing {
             this.marginBeat = marginBeat;
         }
 
-        public int getFees() {
+        public double getFees() {
             return fees;
         }
 
-        public void setFees(int fees) {
+        public void setFees(double fees) {
             this.fees = fees;
         }
